Adicionando Quartz Service e corrigindo código
